// src/contracts/avalon/typechain/OptionMarketViewer.ts
export declare namespace OptionMarketViewer {
// ...
  export type LiquidityBalanceAndAllowanceStructOutput = [
    string,
    BigNumber,
    BigNumber
  ] & { token: string; balance: BigNumber; allowance: BigNumber };
// ...
}