// src/contracts/newport/typechain/OptionMarketViewer.ts
export declare namespace OptionMarketViewer {
// ...
  export type LiquidityBalanceStructOutput = [
    string,
    BigNumber,
    string,
    BigNumber,
    string,
    BigNumber
  ] & {
    quoteAsset: string;
    quoteBalance: BigNumber;
    quoteSymbol: string;
    quoteDepositAllowance: BigNumber;
    liquidityToken: string;
    liquidityBalance: BigNumber;
  };
// ...
}